requirements.txt
# Core dependencies
dbt-core==1.9.1
dbt-postgres==1.9.1

# Data manipulation & analysis
pandas==2.2.3
numpy==2.1.2

# Database connection
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34

# Utilities
python-dotenv==1.0.1
loguru==0.7.2

# Optional: useful for testing and data validation
pytest==8.3.3
great-expectations==1.9.0
